use canvaskit as default and expose CompilationOptions class
